+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 #### Added
 - Added method renameColumn() to migrations
-- Added PHPStorm meta
+- Added PHPStorm meta for better suggestions
 
 ### [1.2.0] - 2020-03-16
 #### Changed